ultimo

+++ Sprint3/laravelXuxemons/database/migrations/2024_04_23_160443_create_chuches_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('chuche_id');
             $table->char('user_id', 6);
-            $table->unsignedBigInteger('stack');
+            $table->integer('stack');
             $table->timestamps();
 
             // Definir las claves foráneas
